requirements.txt
Django==2.1.15
djangorestframework==3.9.4
psycopg2==2.7.7

flake8==3.6.0